requirements.txt
pygame-ce
tkinter
pillow
mediapipe
opencv-python
absl-py